terraform/modules/api/requirements.txt
fastapi
uvicorn
google-cloud-pubsub
google-cloud-bigquery
pydantic

